Add multiple tx msg and move to separate files
