+++ src/components/ExecuteTxDialog/styles.js
@@ -14,6 +14,9 @@ const styles = (theme) => ({
     marginBottom: theme.padding.xs.px,
     overflowX: 'scroll',
   },
+  explanationMsgContainer: {
+    marginBottom: theme.padding.xs.px,
+  },
 });
 
 export default styles;
